Crea archivo para las funciones referidas al paginador
